refactor(stage): split ndi_watchdog.rs into focused sibling modules (#418)
ndi_watchdog.rs was at 996 production lines — 4 from the 1000-line hard cap,
so any further edit would hard-fail CI. Extract the cohesive concerns into
sibling modules following the proven ndi_reload_guard.rs pattern, leaving the
Watchdog lifecycle + page-session ReloadEscalation core in ndi_watchdog.rs:

- ndi_frame_stats.rs (241): FrameStats counters, record_presented_frame,
  present-gap accumulators, the requestVideoFrameCallback observer.
- ndi_profile.rs (155): profile-mode state + localStorage, the bounded
  once-per-pageload switch, proven-mode persistence, maybe_profile_fallback.
- ndi_beacon.rs (184): display id + the /ndi/client-stats stats beacons.
- ndi_health_ticker.rs (124): the 1s health ticker + currentTime frame proxy.

ndi_watchdog.rs drops 996 -> 392 production lines; every resulting file is now
well under the 800-line warning tier. Behavior-preserving mechanical refactor:
each #[cfg(test)] module moved with the code it exercises; cargo check, clippy
(-D warnings), the 121 presenter-ui lib tests, and scripts/dev/quality-check.sh
(0 hard fails) all pass.

Closes #418
