+++ crates/presenter-ui/src/components/stage/mod.rs
@@ -2,7 +2,11 @@ pub mod api_stage;
 pub mod bible_layout;
 pub mod bible_overlay;
 pub mod camera_crew;
+mod ndi_beacon;
+mod ndi_frame_stats;
 pub mod ndi_fullscreen;
+mod ndi_health_ticker;
+mod ndi_profile;
 mod ndi_reload_guard;
 pub mod ndi_video;
 mod ndi_watchdog;
